requirements.txt
click==8.1.7
requests==2.31.0
rich==13.9.4
setuptools==80.9.0
pydantic==2.11.7
packaging==25.0
python-jose==3.5.0